+++ Source/decoder.hpp
@@ -33,9 +33,9 @@ namespace tga
 								   const int height,
 								   color (Decoder::*readPixel)());
 
-		uint8_t read8();
+		//uint8_t read8();
+		int read8();
 		uint16_t read16();
-		//int read16();
 
 		color read8color();
 		color read16AsRgb();
