+++ Source/image_body.hpp
@@ -8,8 +8,8 @@ namespace tga
 		uint8_t* pixels;
 		//uint32_t pixelByteDepth;
 		//uint32_t rowStride;
-		long int pixelByteDepth;
-		long int rowStride;
+		int pixelByteDepth;
+		int rowStride;
 	};
 }
