+++ Source/image_header.hpp
@@ -10,17 +10,28 @@ namespace tga
 {
 	struct ImageHeader
 	{
-		uint8_t idLength;
-		uint8_t colorMapType;
+		//uint8_t idLength;
+		int idLength;
+		//uint8_t colorMapType;
+		int colorMapType;
 		ImageType imageType;
-		int colorMapOrigin;
-		int colorMapLength;
+		//uint16_t colorMapOrigin;
+		uint16_t colorMapOrigin;
+		//uint16_t colorMapLength;
+		uint16_t colorMapLength;
+		//uint8_t colorMapBitDepth;
 		uint8_t colorMapBitDepth;
-		int originX;
-		int originY;
-		int width;
-		int height;
-		uint8_t pixelBitDepth;
+		//uint16_t originX;
+		uint16_t originX;
+		//uint16_t originY;
+		uint16_t originY;
+		//uint16_t width;
+		uint16_t width;
+		//uint16_t height;
+		uint16_t height;
+		//uint8_t pixelBitDepth;
+		int pixelBitDepth;
+		//uint8_t imageDescriptor;
 		uint8_t imageDescriptor;
 		std::string imageId;
 		ColorMap colorMap;
